Improved the AddCustomTagHelperActivation.
Since MVC contains many tag helpers out of the box, custom activation
should not blindly redirect the creation of tag helpers to the
application container, since the application container will not be able
to build up a framework tag helper (with its dependencies). A predicate
argument has been added to the AddCustomTagHelperActivation to allow
control which tag helpers should be built-up by the application
container, and which should be pulled in from the built-in DI system.